+++ src/routes/ConsortiumManager/settings/users/general/PermissionSets/ConsortiumPermissionsSetForm/ConsortiumPermissionsSetForm.test.js
@@ -1,5 +1,5 @@
-import { render, screen, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { render, screen, waitFor } from '@folio/jest-config-stripes/testing-library/react';
+import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import { MemoryRouter, useLocation, useHistory } from 'react-router-dom';
 import {
   QueryClient,