+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 * [UICONSET-113](https://issues.folio.org/browse/UICONSET-113) Allow a user to share settings in the `<ConsortiaControlledVocabulary>`.
 * [UICONSET-49](https://issues.folio.org/browse/UICONSET-49) Display a toast message about lack of permissions for members when loading settings in the consortium members.
 * [UICONSET-136](https://issues.folio.org/browse/UICONSET-136) Replicate assigned users functionality on permission sets.
+* [UICONSET-138](https://issues.folio.org/browse/UICONSET-138) upgrade React to v18